Revert "feat(bbapi): rename bbrpc=>bbapi, move CIVC CLI to use bbapi (#15502)"
This reverts commit 28178392c75a1d0e4c158e2cf703332c9325851e.
