+++ barretenberg/cpp/src/barretenberg/api/api_client_ivc.cpp
@@ -1,7 +1,5 @@
 #include "api_client_ivc.hpp"
-#include "barretenberg/api/bbapi.hpp"
 #include "barretenberg/api/file_io.hpp"
-#include "barretenberg/api/get_bytecode.hpp"
 #include "barretenberg/api/log.hpp"
 #include "barretenberg/api/write_prover_output.hpp"
 #include "barretenberg/client_ivc/client_ivc.hpp"
@@ -15,11 +13,37 @@
 #include "barretenberg/serialize/msgpack.hpp"
 #include "barretenberg/serialize/msgpack_check_eq.hpp"
 #include <algorithm>
-#include <sstream>
 #include <stdexcept>
 
 namespace bb {
-namespace { // anonymous namespace
+
+acir_format::WitnessVector witness_map_to_witness_vector(std::map<std::string, std::string> const& witness_map)
+{
+    acir_format::WitnessVector wv;
+    size_t index = 0;
+    for (auto& e : witness_map) {
+        uint64_t value = stoull(e.first);
+        // ACIR uses a sparse format for WitnessMap where unused witness indices may be left unassigned.
+        // To ensure that witnesses sit at the correct indices in the `WitnessVector`, we fill any indices
+        // which do not exist within the `WitnessMap` with the dummy value of zero.
+        while (index < value) {
+            wv.push_back(fr(0));
+            index++;
+        }
+        wv.push_back(fr(uint256_t(e.second)));
+        index++;
+    }
+    return wv;
+}
+
+std::shared_ptr<ClientIVC::DeciderProvingKey> get_acir_program_decider_proving_key(acir_format::AcirProgram& program)
+{
+    ClientIVC::ClientCircuit builder = acir_format::create_circuit<ClientIVC::ClientCircuit>(program);
+
+    // Construct the verification key via the prover-constructed proving key with the proper trace settings
+    TraceSettings trace_settings{ AZTEC_TRACE_STRUCTURE };
+    return std::make_shared<ClientIVC::DeciderProvingKey>(builder, trace_settings);
+}
 
 /**
  * @brief Compute and write to file a MegaHonk VK for a circuit to be accumulated in the IVC
@@ -29,23 +53,24 @@ namespace { // anonymous namespace
  * @param bytecode_path
  * @param witness_path
  */
-void write_standalone_vk(const std::string& output_format,
-                         const std::filesystem::path& bytecode_path,
-                         const std::filesystem::path& output_path)
+void write_standalone_vk(const std::string& output_data_type,
+                         const std::string& bytecode_path,
+                         const std::string& output_path)
 {
-    auto bytecode = get_bytecode(bytecode_path);
-    auto response = bbapi::ClientIvcComputeStandaloneVk{
-        .circuit = { .name = "standalone_circuit", .bytecode = std::move(bytecode) }
-    }.execute();
 
-    if (output_format == "bytes") {
-        write_file(output_path / "vk", response.bytes);
-    } else if (output_format == "fields") {
-        std::string json = field_elements_to_json(response.fields);
-        write_file(output_path / "vk_fields.json", std::vector<uint8_t>(json.begin(), json.end()));
-    } else {
-        throw_or_abort("Unsupported output format for standalone vk: " + output_format);
-    }
+    acir_format::AcirProgram program{ get_constraint_system(bytecode_path), /*witness=*/{} };
+    std::shared_ptr<ClientIVC::DeciderProvingKey> proving_key = get_acir_program_decider_proving_key(program);
+    auto verification_key = std::make_shared<ClientIVC::MegaVerificationKey>(proving_key->get_precomputed());
+    PubInputsProofAndKey<ClientIVC::MegaVerificationKey> to_write{ .key = verification_key };
+
+    write(to_write, output_data_type, "vk", output_path);
+}
+
+size_t get_num_public_inputs_in_circuit(const std::filesystem::path& bytecode_path)
+{
+    using namespace acir_format;
+    acir_format::AcirProgram program{ get_constraint_system(bytecode_path), /*witness=*/{} };
+    return program.constraints.public_inputs.size();
 }
 
 void write_vk_for_ivc(const std::string& output_format,
@@ -55,15 +80,25 @@ void write_vk_for_ivc(const std::string& output_format,
     if (output_format != "bytes") {
         throw_or_abort("Unsupported output format for ClientIVC vk: " + output_format);
     }
+    ClientIVC ivc{ { AZTEC_TRACE_STRUCTURE } };
+    ClientIVCMockCircuitProducer circuit_producer;
 
-    // Since we need to specify the number of public inputs but ClientIvcComputeIvcVk derives it from bytecode,
-    // we need to create a mock circuit with the correct number of public inputs
-    // For now, we'll use the compute_vk_for_ivc function directly as it was designed for this purpose
-    bbapi::BBApiRequest request;
-    auto vk = bbapi::compute_vk_for_ivc(request, num_public_inputs_in_final_circuit);
-    const auto buf = to_buffer(vk);
+    // Initialize the IVC with an arbitrary circuit
+    // We segfault if we only call accumulate once
+    static constexpr size_t SMALL_ARBITRARY_LOG_CIRCUIT_SIZE{ 5 };
+    MegaCircuitBuilder circuit_0 = circuit_producer.create_next_circuit(ivc, SMALL_ARBITRARY_LOG_CIRCUIT_SIZE);
+    ivc.accumulate(circuit_0);
+
+    // Create another circuit and accumulate
+    MegaCircuitBuilder circuit_1 =
+        circuit_producer.create_next_circuit(ivc, SMALL_ARBITRARY_LOG_CIRCUIT_SIZE, num_public_inputs_in_final_circuit);
+    ivc.accumulate(circuit_1);
+
+    // Construct the hiding circuit and its VK (stored internally in the IVC)
+    ivc.construct_hiding_circuit_key();
 
     const bool output_to_stdout = output_dir == "-";
+    const auto buf = to_buffer(ivc.get_vk());
 
     if (output_to_stdout) {
         write_bytes_to_stdout(buf);
@@ -76,47 +111,28 @@ void write_vk_for_ivc(const std::string& output_data_type,
                       const std::string& bytecode_path,
                       const std::filesystem::path& output_dir)
 {
-    if (output_data_type != "bytes") {
-        throw_or_abort("Unsupported output format for ClientIVC vk: " + output_data_type);
-    }
-
-    auto bytecode = get_bytecode(bytecode_path);
-
-    auto response = bbapi::ClientIvcComputeIvcVk{
-        .circuit = { .name = "final_circuit", .bytecode = std::move(bytecode) }
-    }.execute();
-
-    const bool output_to_stdout = output_dir == "-";
-    if (output_to_stdout) {
-        write_bytes_to_stdout(response.bytes);
-    } else {
-        write_file(output_dir / "vk", response.bytes);
-    }
+    const size_t num_public_inputs_in_final_circuit = get_num_public_inputs_in_circuit(bytecode_path);
+    info("num_public_inputs_in_final_circuit: ", num_public_inputs_in_final_circuit);
+    write_vk_for_ivc(output_data_type, num_public_inputs_in_final_circuit, output_dir);
 }
-} // anonymous namespace
 
 void ClientIVCAPI::prove(const Flags& flags,
                          const std::filesystem::path& input_path,
                          const std::filesystem::path& output_dir)
 {
 
-    bbapi::BBApiRequest request;
+    PrivateExecutionSteps steps;
+    steps.parse(PrivateExecutionStepRaw::load_and_decompress(input_path));
 
-    bbapi::ClientIvcStart{}.execute(request);
-    std::vector<PrivateExecutionStepRaw> raw_steps = PrivateExecutionStepRaw::load_and_decompress(input_path);
+    std::shared_ptr<ClientIVC> ivc = steps.accumulate();
+    ClientIVC::Proof proof = ivc->prove();
 
-    size_t last_circuit_public_inputs_size = 0;
-    for (const auto& step : raw_steps) {
-        bbapi::ClientIvcLoad{
-            .circuit = { .name = step.function_name, .bytecode = step.bytecode, .verification_key = step.vk }
-        }.execute(request);
-
-        last_circuit_public_inputs_size = request.last_circuit_constraints->public_inputs.size();
-        info("ClientIVC: accumulating " + step.function_name);
-        bbapi::ClientIvcAccumulate{ .witness = step.witness }.execute(request);
+    // We verify this proof. Another bb call to verify has the overhead of loading the SRS,
+    // and it is mysterious if this transaction fails later in the lifecycle.
+    // The files are still written in case they are needed to investigate this failure.
+    if (!ivc->verify(proof)) {
+        THROW std::runtime_error("Failed to verify the private (ClientIVC) transaction proof!");
     }
-
-    auto proof = bbapi::ClientIvcProve{}.execute(request).proof;
 
     // We'd like to use the `write` function that UltraHonkAPI uses, but there are missing functions for creating
     // std::string representations of vks that don't feel worth implementing
@@ -137,7 +153,8 @@ void ClientIVCAPI::prove(const Flags& flags,
 
     if (flags.write_vk) {
         vinfo("writing ClientIVC vk in directory ", output_dir);
-        write_vk_for_ivc("bytes", last_circuit_public_inputs_size, output_dir);
+        const size_t num_public_inputs_in_final_circuit = steps.folding_stack.back().constraints.public_inputs.size();
+        write_vk_for_ivc("bytes", num_public_inputs_in_final_circuit, output_dir);
     }
 }
 
@@ -177,22 +194,21 @@ void ClientIVCAPI::write_solidity_verifier([[maybe_unused]] const Flags& flags,
 }
 
 bool ClientIVCAPI::check_precomputed_vks(const std::filesystem::path& input_path)
-{
-    bbapi::BBApiRequest request;
-    std::vector<PrivateExecutionStepRaw> raw_steps = PrivateExecutionStepRaw::load_and_decompress(input_path);
 
-    for (const auto& step : raw_steps) {
-        if (step.vk.empty()) {
-            info("FAIL: Expected precomputed vk for function ", step.function_name);
+{
+    PrivateExecutionSteps steps;
+    steps.parse(PrivateExecutionStepRaw::load_and_decompress(input_path));
+
+    for (auto [program, precomputed_vk, function_name] :
+         zip_view(steps.folding_stack, steps.precomputed_vks, steps.function_names)) {
+        if (precomputed_vk == nullptr) {
+            info("FAIL: Expected precomputed vk for function ", function_name);
             return false;
         }
-        auto response = bbapi::ClientIvcCheckPrecomputedVk{ .circuit = { .name = step.function_name,
-                                                                         .bytecode = step.bytecode,
-                                                                         .verification_key = step.vk },
-                                                            .function_name = step.function_name }
-                            .execute();
-
-        if (!response.valid) {
+        std::shared_ptr<ClientIVC::DeciderProvingKey> proving_key = get_acir_program_decider_proving_key(program);
+        auto computed_vk = std::make_shared<ClientIVC::MegaVerificationKey>(proving_key->get_precomputed());
+        std::string error_message = "FAIL: Precomputed vk does not match computed vk for function " + function_name;
+        if (!msgpack::msgpack_check_eq(*computed_vk, *precomputed_vk, error_message)) {
             return false;
         }
     }
