+++ barretenberg/cpp/src/barretenberg/api/api_client_ivc.hpp
@@ -44,6 +44,8 @@ void gate_count_for_ivc(const std::string& bytecode_path, bool include_gates_per
 
 void write_arbitrary_valid_client_ivc_proof_and_vk_to_file(const std::filesystem::path& output_dir);
 
+acir_format::WitnessVector witness_map_to_witness_vector(std::map<std::string, std::string> const& witness_map);
+
 std::vector<uint8_t> decompress(const void* bytes, size_t size);
 
 } // namespace bb
