--- barretenberg/cpp/src/barretenberg/api/bbapi.hpp
@@ -1,13 +1,0 @@
-#pragma once
-/**
- * @file bbapi_commands.hpp
- * @brief Central command definitions BB's outward-facing API.
- *
- * This file includes and exports all command structures from specialized headers
- * and provides unified Command and CommandResponse types for the API.
- */
-#include "barretenberg/api/bbapi_client_ivc.hpp"
-#include "barretenberg/api/bbapi_execute.hpp"
-#include "barretenberg/api/bbapi_shared.hpp"
-#include "barretenberg/api/bbapi_ultra_honk.hpp"
-#include "barretenberg/common/named_union.hpp"
