+++ ci3/aws_request_instance
@@ -47,10 +47,6 @@ else
   exit 1
 fi
 
-if [ -n "${AWS_INSTANCE:-}" ]; then
-  instances=("$AWS_INSTANCE")
-fi
-
 set +e
 IFS=',' read -ra cpu_list <<< "$cpus"
 for cpu in "${cpu_list[@]}"; do
